requirements.txt
Django==3.2.12
djangorestframework==3.12.4
mysqlclient==2.0.3
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
Pillow==8.4.0
psycopg2-binary==2.9.3
django-filter==2.4.0
django-cors-headers==3.7.0